+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.9dev
 * [Feature] Adds `--no-tag` option to `pkgmt version`
+* [Feature] Adds `--target stable` option to `pkgmt version`
 
 ## 0.1.8 (2023-02-09)